add sqlite plugin

+++ sdkproject/Assets/Mapbox/Core/mapbox-sdk-cs/Platform/SQLite/SQLiteDataService.cs
@@ -23,44 +23,54 @@ namespace Mapbox.Platform.SQLite
 #if UNITY_EDITOR
 			var dbPath = string.Format(@"Assets/StreamingAssets/{0}", DatabaseName);
 #else
-        // check if file exists in Application.persistentDataPath
-        var filepath = string.Format("{0}/{1}", Application.persistentDataPath, DatabaseName);
+			// check if file exists in Application.persistentDataPath
+			var filepath = string.Format("{0}/{1}", Application.persistentDataPath, DatabaseName);
 
-        if (!File.Exists(filepath))
-        {
-            Debug.Log("Database not in Persistent path");
-            // if it doesn't ->
-            // open StreamingAssets directory and load the db ->
+			Debug.LogErrorFormat("filepath: {0}", filepath);
+
+			if (!File.Exists(filepath))
+			{
+				Debug.Log("Database not in Persistent path");
+				// if it doesn't ->
+				// open StreamingAssets directory and load the db ->
 
 #if UNITY_ANDROID
-            var loadDb = new WWW("jar:file://" + Application.dataPath + "!/assets/" + DatabaseName);  // this is the path to your StreamingAssets in android
-            while (!loadDb.isDone) { }  // CAREFUL here, for safety reasons you shouldn't let this while loop unattended, place a timer and error check
-            // then save to Application.persistentDataPath
-            File.WriteAllBytes(filepath, loadDb.bytes);
+				var loadDb = new WWW("jar:file://" + Application.dataPath + "!/assets/" + DatabaseName);  // this is the path to your StreamingAssets in android
+				while (!loadDb.isDone) { }  // CAREFUL here, for safety reasons you shouldn't let this while loop unattended, place a timer and error check
+				// then save to Application.persistentDataPath
+				File.WriteAllBytes(filepath, loadDb.bytes);
 #elif UNITY_IOS
-                 var loadDb = Application.dataPath + "/Raw/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
-                // then save to Application.persistentDataPath
-                File.Copy(loadDb, filepath);
+				var loadDb = Application.dataPath + "/Raw/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
+				// then save to Application.persistentDataPath
+				File.Copy(loadDb, filepath);
 #elif UNITY_WP8
-                var loadDb = Application.dataPath + "/StreamingAssets/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
-                // then save to Application.persistentDataPath
-                File.Copy(loadDb, filepath);
+				var loadDb = Application.dataPath + "/StreamingAssets/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
+				// then save to Application.persistentDataPath
+				File.Copy(loadDb, filepath);
 
 #elif UNITY_WINRT
-		var loadDb = Application.dataPath + "/StreamingAssets/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
-		// then save to Application.persistentDataPath
-		File.Copy(loadDb, filepath);
+				var loadDb = Application.dataPath + "/StreamingAssets/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
+				// then save to Application.persistentDataPath
+				if (File.Exists(loadDb))
+				{
+					File.Copy(loadDb, filepath);
+				}
 #else
-	var loadDb = Application.dataPath + "/StreamingAssets/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
-	// then save to Application.persistentDataPath
-	File.Copy(loadDb, filepath);
-
+				var loadDb = Application.dataPath + "/StreamingAssets/" + DatabaseName;  // this is the path to your StreamingAssets in iOS
+																						 // then save to Application.persistentDataPath
+				// only copy if db exists
+				if (File.Exists(loadDb))
+				{
+					File.Copy(loadDb, filepath);
+				}
+				Debug.LogErrorFormat("loadDb: {0}", loadDb);
+				Debug.LogErrorFormat("filepath: {0}", filepath);
 #endif
 
-            Debug.Log("Database written");
-        }
+				Debug.Log("Database written");
+			}
 
-        var dbPath = filepath;
+			var dbPath = filepath;
 #endif
 			_connection = new SQLiteConnection(dbPath, SQLiteOpenFlags.ReadWrite | SQLiteOpenFlags.Create);
 			Debug.Log("Final PATH: " + dbPath);
@@ -118,9 +128,9 @@ namespace Mapbox.Platform.SQLite
 		}
 
 
-		public void CreateTable<T>()
+		public int CreateTable<T>()
 		{
-			_connection.CreateTable<T>();
+			return _connection.CreateTable<T>();
 		}
